FPLA-1324 renamed demo to adoption
