+++ src/integrationTest/java/uk/gov/hmcts/reform/adoption/controllers/GetWelcomeTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.controllers;
+package uk.gov.hmcts.reform.adoption.controllers;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
